refactor: schedule_detail_page 720줄 → 4파일 분리
- schedule_formatters.dart: 날짜/상태/RRULE 순수 로직 유틸 추출
- info_row.dart: 범용 아이콘+라벨+값 위젯 분리
- schedule_timer_section.dart: 자체 provider watch하는 독립 섹션
- calendar_view.dart: 중복 _isSameDay 제거, 공유 유틸 사용
